+++ tests/fixtures/application_fixtures.py
@@ -42,6 +42,7 @@ from fides.api.ops.models.privacy_experience import (
     ComponentType,
     DeliveryMechanism,
     PrivacyExperience,
+    PrivacyExperienceConfig,
 )
 from fides.api.ops.models.privacy_notice import (
     ConsentMechanism,
@@ -2048,52 +2049,92 @@ def consent_records(
 
 
 @pytest.fixture(scope="function")
-def privacy_experience_privacy_center_link(db: Session) -> Generator:
+def experience_config_privacy_center(db: Session) -> Generator:
+    exp = PrivacyExperienceConfig.create(
+        db=db,
+        data={
+            "component": "privacy_center",
+            "delivery_mechanism": "link",
+            "component_title": "Control your privacy",
+            "link_label": "Manage your preferences",
+        },
+    )
+    yield exp
+    for history in exp.histories:
+        history.delete(db)
+    exp.delete(db)
+
+
+@pytest.fixture(scope="function")
+def privacy_experience_privacy_center_link(
+    db: Session, experience_config_privacy_center
+) -> Generator:
     privacy_experience = PrivacyExperience.create(
         db=db,
         data={
             "component": ComponentType.privacy_center,
+            "disabled": False,
             "delivery_mechanism": DeliveryMechanism.link,
-            "regions": [
-                PrivacyNoticeRegion.us_ca,
-                PrivacyNoticeRegion.us_co,
-            ],
+            "region": PrivacyNoticeRegion.us_co,
+            "experience_config_id": experience_config_privacy_center.id,
+            "experience_config_history_id": experience_config_privacy_center.histories[
+                0
+            ].id,
+        },
+    )
+
+    yield privacy_experience
+    for history in privacy_experience.histories:
+        history.delete(db)
+    privacy_experience.delete(db)
+
+
+@pytest.fixture(scope="function")
+def experience_config_overlay_link(db: Session) -> Generator:
+    exp = PrivacyExperienceConfig.create(
+        db=db,
+        data={
+            "component": "overlay",
+            "delivery_mechanism": "link",
             "component_title": "Manage your consent preferences",
             "component_description": "On this page you can opt in and out of these data uses cases",
             "link_label": "Manage your privacy",
         },
     )
-
-    yield privacy_experience
+    yield exp
+    for history in exp.histories:
+        history.delete(db)
+    exp.delete(db)
 
 
 @pytest.fixture(scope="function")
-def privacy_experience_overlay_link(db: Session) -> Generator:
+def privacy_experience_overlay_link(
+    db: Session, experience_config_overlay_link
+) -> Generator:
     privacy_experience = PrivacyExperience.create(
         db=db,
         data={
             "component": ComponentType.overlay,
             "delivery_mechanism": DeliveryMechanism.link,
-            "regions": [PrivacyNoticeRegion.eu_fr],
-            "component_title": "Manage your consent preferences",
-            "component_description": "On this page you can opt in and out of these data uses cases",
-            "link_label": "Manage your privacy",
+            "region": PrivacyNoticeRegion.eu_fr,
+            "experience_config_id": experience_config_overlay_link.id,
+            "experience_config_history_id": experience_config_overlay_link.histories[
+                0
+            ].id,
         },
     )
 
     yield privacy_experience
+    for history in privacy_experience.histories:
+        history.delete(db)
+    privacy_experience.delete(db)
 
 
 @pytest.fixture(scope="function")
-def privacy_experience_overlay_banner(db: Session) -> Generator:
-    privacy_experience = PrivacyExperience.create(
+def experience_config_overlay_banner(db: Session) -> Generator:
+    exp = PrivacyExperienceConfig.create(
         db=db,
         data={
-            "component": ComponentType.overlay,
-            "delivery_mechanism": DeliveryMechanism.banner,
-            "regions": [
-                PrivacyNoticeRegion.us_ca,
-            ],
             "component_title": "Manage your consent",
             "component_description": "On this page you can opt in and out of these data uses cases",
             "banner_title": "Manage your consent",
@@ -2101,7 +2142,36 @@ def privacy_experience_overlay_banner(db: Session) -> Generator:
             "confirmation_button_label": "Accept all",
             "reject_button_label": "Reject all",
             "disabled": True,
+            "component": ComponentType.overlay,
+            "delivery_mechanism": DeliveryMechanism.banner,
+            "acknowledgement_button_label": "Confirm",
+        },
+    )
+    yield exp
+    for history in exp.histories:
+        history.delete(db)
+    exp.delete(db)
+
+
+@pytest.fixture(scope="function")
+def privacy_experience_overlay_banner(
+    db: Session, experience_config_overlay_banner
+) -> Generator:
+    privacy_experience = PrivacyExperience.create(
+        db=db,
+        data={
+            "component": ComponentType.overlay,
+            "delivery_mechanism": DeliveryMechanism.banner,
+            "region": PrivacyNoticeRegion.us_ca,
+            "experience_config_id": experience_config_overlay_banner.id,
+            "experience_config_history_id": experience_config_overlay_banner.histories[
+                0
+            ].id,
+            "disabled": True,
         },
     )
 
     yield privacy_experience
+    for history in privacy_experience.histories:
+        history.delete(db)
+    privacy_experience.delete(db)