+++ tests/ops/api/v1/test_utils.py
@@ -5,6 +5,7 @@ import pytest
 from fastapi import HTTPException
 
 from fides.api.ops.api.v1.endpoints.utils import (
+    human_friendly_list,
     transform_fields,
     validate_start_and_end_filters,
 )
@@ -42,3 +43,31 @@ class TestTransformFields:
         notice_2 = transform_fields(escape, notice_2, ["description"])
 
         assert notice_1 == notice_2
+
+
+class TestHumanFriendlyList:
+    def test_no_items(self):
+        assert human_friendly_list([]) == ""
+        assert human_friendly_list(None) == ""
+
+    def test_one_item(self):
+        assert human_friendly_list(["red"]) == "red"
+        assert human_friendly_list([1]) == "1"
+
+    def test_two_items(self):
+        assert human_friendly_list(["red", "blue"]) == "red and blue"
+        assert human_friendly_list([1, "blue"]) == "1 and blue"
+
+    def test_three_items(self):
+        assert human_friendly_list(["red", "blue", "yellow"]) == "red, blue, and yellow"
+        assert human_friendly_list([1, "blue", "yellow"]) == "1, blue, and yellow"
+
+    def test_four_items(self):
+        assert (
+            human_friendly_list(["red", "blue", "yellow", "green"])
+            == "red, blue, yellow, and green"
+        )
+        assert (
+            human_friendly_list([1, "blue", "yellow", "green"])
+            == "1, blue, yellow, and green"
+        )